Add files via upload
Testing hello world options. The script allows the user to speak to the robot unit and say hello or good morning (does not work on a virtual robot) and the robot gives a response. Also created other options to program text-to-speech via Python and using the "say" box in the program.